feat: refactor PongGame initialization and add AI battle button functionality
