+++ sample-3D/src/3Dpong/PongGame.ts
@@ -1,8 +1,10 @@
+import { Engine } from "@babylonjs/core";
 import { PongGUI } from "../gui/PongGUI";
 import { ScoreBoardGUI } from "../gui/ScoreBoardGUI";
 import { Pong } from "./Pong";
 import { PongSender } from "./PongSender";
 import { PongUpdater } from "./PongUpdater";
+import { PongBuilder } from "./PongBuilder";
 
 
 export class PongGame {
@@ -13,6 +15,34 @@ export class PongGame {
 		scoreboard: ScoreBoardGUI,
 	}) {
 
+	}
+
+	// PongGameの作成
+	public static async bootPongGame(canvas: HTMLCanvasElement)
+	{
+		const engine = new Engine(canvas, true); // Generate the BABYLON 3D engine
+		// ゲーム画面とGUIの初期化
+		const pong = PongBuilder.CreatePong(engine, canvas);
+		const pongGui = await PongGUI.createPongGUI(pong.props.scene);
+		pongGui.setScore(0, 0);
+		const scoreboard = await ScoreBoardGUI.createScoreBoardGUI(pong.props.scene);
+
+		// 継続的にシーンをレンダリングする
+		engine.runRenderLoop(function () {
+			pong.props.scene.render();
+		});
+		
+		// ウィンドウのリサイズイベントを監視して、エンジンをリサイズ
+		window.addEventListener("resize", function () {
+			engine.resize();
+		});
+
+		const pongGame = new PongGame({
+			pong,
+			pongGui,
+			scoreboard,
+		})
+		return pongGame;
 	}
 
 	public async createAiGame(props: {
